Cache fixes
The datation was incorrect on the server-side, resulting in the data being fetched everytime.
On the client, asynchronousness is better handled, so that we don't get another countries's stereotype for the country we are hovering when the network is slow.
Also a bug with position: fixed on chrome mobile is now worked around.